requirements.txt
streamlit
openai
pypdf
pandas
fpdf2